requirements.txt
flask==3.1.*
gunicorn==23.*
pyyaml==6.*
psutil==6.*
paho-mqtt==2.*
